Draw the sole selected unit even if not in viewport
If only one unit is selected, then draw that one even if it is not in
the viewport, as long as the player is entitled to see it.  This is
because the sight, react, and attack circles drawn by DrawInformations
may still reach the viewport.

This gets a bit messy because an enemy unit hidden in the fog of war
must not be drawn even if selected.  Still, all the functions involved
are defined in the same file, so this hack is not too awful.

This fixes bug #27629 "minor UI issue - only units in viewed area get drawn".

Patch by Kalle Olavi Niemitalo.