Base the CanTarget checks on MovementMask
The shipyard is a naval unit but also allowed on land.  It is not a
ShoreBuilding because there aren't enough coast tiles in the Gosh
patches.  This combination made it attackable only by units that have
CanTargetSea=true; CanTargetLand had no effect even though the
shipyard typically has 5 or so land tiles under it.

To make the shipyard attackable also by units that have
CanTargetLand=true, change the CanTarget function so that it treats
UnitTypeLand and UnitTypeNaval as equivalent, and instead examines
MovementMask.  With the Bos Wars game data, the shipyard is the only
unit type affected by this change, I think.