After a unit is fully harvested, look for similar ones
When a worker finishes harvesting and has no other orders, it looks
for another harvestable unit.  Change this so that if the harvested
unit (e.g. a tree or a rock) contained only one type of resource
(energy or magma), then the worker looks only for units that contain
the same type of resource.  This should help the income stay balanced.

Fixes: http://savannah.nongnu.org/bugs/?30856
 (Workers swich to other resourse type after one deposit complete)