Remove documentation of DecorationOnTop
CclDecorationOnTop was never implemented for Lua and was removed in
bos SVN trunk r8614 on 2007-03-15.  And I doubt it worked with Guile
either.

I've filed bug #29097 on finally implementing a similar feature
but DecorationOnTop as documented seems the wrong way to do it;
if the feature can be enabled at run time, there should surely
be a way to disable it again as well.