gcn::DefaultFont::drawString: Correctly increment x
This was doing x += getWidth(text) after drawing each glyph,
but getWidth returns the width of the whole string, thus being
wrong if the string contains more than one glyph.
drawGlyph has already been documented as returning the width
of the single glyph, so use that instead.

This fix is not particularly important, because gcn::DefaultFont
is only a fallback for when the correct font has not been set.