Don't recognize tilde syntaxes in text input fields
If the user types something like "foo~!bar" in a TextField widget,
then display the tilde as is; the "~!" there does not mean that the
next character is a hot key.  The user is not supposed to know about
any such syntax.

Implemented by having a parallel plain-text CFont that Lua code then
assigns to each TextField it creates.  The plain-text CFont shares
textures with the markup CFont, so the memory cost is minimal.

(Originally, I was instead going to implement this by adding a syntax
parameter to gcn::Font::drawString and all related functions.  That
however was the wrong choice because e.g. gcn::ImageFont::drawString
then had to take the parameter but had no idea what to do with it.)

Plain-text fonts can be used in all other widgets too, such as list
boxes or labels.  This commit does not make any such widgets use them
though, and the choice of font does not currently affect how the
engine checks whether the widget has a hot key defined.