make.py: fix Check(function=)
Check(b, function='strcasestr') always returned True because it only
declared strcasestr() and the linker never checked whether the
function actually existed.  Make it also call the function so that
the linker returns an error if the function does not exist.
The arguments for the call do not have to make sense because the
resulting executable will not be executed.

Now that Check properly detects the absence of strcasestr and strnlen,
the explicit -UHAVE_STRCASESTR -UHAVE_STRNLEN in the mingw
configuration should be unnecessary.  However, I haven't properly
tested building that configuration, so I'm leaving them in for now.