doxygen INPUT_FILTER: search for tabs too
In tools/doxygen-boswars.cfg, INPUT_FILTER is a long sed command that
converts comments in Bos Wars source files to something that Doxygen
understands.  In particular, on lines like

CFont *GameFont;        /// Normal font used in game

it changes the "///" to "//!<" so that Doxygen attaches the comment
to the previous definition rather than to the next one.

On the other hand, there are also many pairs of lines like

	/// Real font width (starting with ' ')
	char *CharWidth;

where the comment should attach to the definition on the next line.
In the INPUT_FILTER, the 's:^[  ]*///://\!:' sed command was apparently
intended to recognize these cases.  However, because the two characters
between the brackets were both spaces, this command did not match lines
that were indented with tabs.  This caused Doxygen to attach the comments
to the wrong declarations.  Change the first space to a tab character,
so that the regex will match any mixture of tabs and spaces.